requirements.dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.dev.txt requirements.dev.in
#
attrs==23.1.0
    # via
    #   cattrs
    #   fiona
    #   requests-cache
black==23.9.1
    # via -r requirements.dev.in
build==0.10.0
    # via pip-tools
cattrs==23.1.2
    # via requests-cache
certifi==2023.7.22
    # via requests
cfgv==3.3.1
    # via pre-commit
charset-normalizer==3.2.0
    # via requests
click==8.1.6
    # via
    #   black
    #   click-plugins
    #   cligj
    #   fiona
    #   pip-tools
click-plugins==1.1.1
    # via fiona
cligj==0.7.2
    # via fiona
contourpy==1.1.0
    # via matplotlib
country-converter==1.0.0
    # via -r requirements.dev.in
coverage==7.3.1
    # via -r requirements.dev.in
cycler==0.11.0
    # via matplotlib
distlib==0.3.7
    # via virtualenv
filelock==3.12.2
    # via virtualenv
fiona==1.9.4.post1
    # via geopandas
fonttools==4.42.1
    # via matplotlib
geopandas==0.13.2
    # via -r requirements.dev.in
identify==2.5.25
    # via pre-commit
idna==3.4
    # via
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
kiwisolver==1.4.5
    # via matplotlib
matplotlib==3.7.2
    # via -r requirements.dev.in
multidict==6.0.4
    # via yarl
mypy-extensions==1.0.0
    # via black
nodeenv==1.8.0
    # via pre-commit
numpy==1.26.0
    # via
    #   -r requirements.dev.in
    #   contourpy
    #   matplotlib
    #   pandas
packaging==23.0
    # via
    #   black
    #   build
    #   geopandas
    #   matplotlib
    #   pytest
pandas==2.1.0
    # via
    #   -r requirements.dev.in
    #   country-converter
pathspec==0.9.0
    # via black
pillow==10.0.0
    # via matplotlib
pip-tools==7.2.0
    # via -r requirements.dev.in
platformdirs==3.10.0
    # via
    #   black
    #   requests-cache
    #   virtualenv
pluggy==1.2.0
    # via pytest
pre-commit==3.4.0
    # via -r requirements.dev.in
pyparsing==3.0.9
    # via matplotlib
pyproj==3.6.0
    # via geopandas
pyproject-hooks==1.0.0
    # via build
pytest==7.4.2
    # via -r requirements.dev.in
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2023.3
    # via pandas
pyyaml==6.0.1
    # via
    #   pre-commit
    #   vcrpy
requests==2.31.0
    # via
    #   -r requirements.dev.in
    #   requests-cache
requests-cache==1.1.0
    # via -r requirements.dev.in
ruff==0.0.290
    # via -r requirements.dev.in
schemdraw==0.17
    # via -r requirements.dev.in
shapely==2.0.1
    # via geopandas
six==1.16.0
    # via
    #   fiona
    #   python-dateutil
    #   url-normalize
tomli==1.1.0
    # via
    #   black
    #   build
    #   pip-tools
    #   pyproject-hooks
    #   pytest
typing-extensions==4.7.1
    # via cattrs
tzdata==2023.3
    # via pandas
unidecode==1.3.6
    # via -r requirements.dev.in
url-normalize==1.4.3
    # via requests-cache
urllib3==2.0.4
    # via
    #   requests
    #   requests-cache
vcrpy==5.1.0
    # via -r requirements.dev.in
virtualenv==20.24.2
    # via pre-commit
wheel==0.41.0
    # via pip-tools
wrapt==1.15.0
    # via vcrpy
yarl==1.9.2
    # via vcrpy

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
